Convert `src/redux/showcaseTokens` to TypeScript.
